+++ geti_sdk/rest_clients/deployment_client.py
@@ -14,23 +14,21 @@
 import logging
 import os
 import tempfile
-import time
 import zipfile
 from typing import List, Optional, Sequence, Union
 
 from geti_sdk.data_models import Project
 from geti_sdk.data_models.code_deployment_info import (
-    CodeDeploymentInformation,
     DeploymentModelIdentifier,
 )
-from geti_sdk.data_models.enums import DeploymentState, OptimizationType
+from geti_sdk.data_models.enums import OptimizationType
 from geti_sdk.data_models.model import Model, OptimizedModel
 from geti_sdk.deployment import Deployment
 from geti_sdk.http_session import GetiSession
 from geti_sdk.rest_clients.configuration_client import ConfigurationClient
 from geti_sdk.rest_clients.model_client import ModelClient
 from geti_sdk.rest_clients.prediction_client import PredictionClient
-from geti_sdk.utils import deserialize_dictionary, get_supported_algorithms
+from geti_sdk.utils import get_supported_algorithms
 
 
 class DeploymentClient:
@@ -58,13 +56,13 @@ class DeploymentClient:
         )
 
     @property
-    def code_deployment_url(self) -> str:
+    def deployment_package_url(self) -> str:
         """
-        Return the base URL for the code deployment group of endpoints
+        Return the base URL for the deployment group of endpoints
 
-        :return: URL for the code deployment endpoints for the Intel® Geti™ project
+        :return: URL for the deployment endpoints for the Intel® Geti™ project
         """
-        return self.base_url + "/code_deployments"
+        return self.base_url + "/deployment_package"
 
     @property
     def ready_to_deploy(self) -> bool:
@@ -80,63 +78,24 @@ class DeploymentClient:
             model is not None for model in self._model_client.get_all_active_models()
         )
 
-    def _request_deployment(
-        self, model_identifiers: Sequence[DeploymentModelIdentifier]
-    ) -> CodeDeploymentInformation:
+    def _fetch_deployment(
+        self, model_identifiers: list[DeploymentModelIdentifier]
+    ) -> Deployment:
         """
-        Request the creation of a code deployment for the project, using the models
-        with identifiers as passed in `models`
+        Download and return the deployment.
 
-        :param model_identifiers: List of identifiers for the models to use in the
-            deployment
-        :return: CodeDeploymentInformation object, holding all available information
-            regarding the state of the deployment creation process on the Intel® Geti™
-            server.
-        """
-        model_id_data = [model.to_dict() for model in model_identifiers]
-        response = self.session.get_rest_response(
-            url=self.code_deployment_url + ":prepare",
-            method="POST",
-            data={"models": model_id_data},
-        )
-        logging.info(f"Deployment for project '{self.project.name}' started.")
-        return deserialize_dictionary(response, CodeDeploymentInformation)
-
-    def _get_deployment_status(self, deployment_id: str) -> CodeDeploymentInformation:
-        """
-        Return the actual status of a code deployment creation process on the Intel®
-        Geti™ server.
-
-        :param deployment_id: Unique identifier of the deployment to get the status for
-        :return: CodeDeploymentInformation instance holding the current state of the
-            deployment creation process
-        """
-        response = self.session.get_rest_response(
-            url=self.code_deployment_url + f"/{deployment_id}", method="GET"
-        )
-        return deserialize_dictionary(response, CodeDeploymentInformation)
-
-    def _fetch_deployment(self, deployment_id: str) -> Deployment:
-        """
-        Download and return the deployment identified by `deployment_id`. If the
-        deployment is not ready to be downloaded yet, this method will raise an error.
-
-        :param deployment_id: Unique identifier of the deployment to download
+        :param model_identifiers: list of models to include in the deployment
         :return: `Deployment` object containing the deployment
         """
-        status = self._get_deployment_status(deployment_id=deployment_id)
-        if status.state != DeploymentState.DONE:
-            raise ValueError(
-                f"Deployment with ID '{deployment_id}' is not ready to be retrieved "
-                f"(yet). The current state is '{status.state}'. Deployment can only be "
-                f"retrieved when it is in state 'DONE'. "
-            )
+
+        data = {
+            "package_type": "geti_sdk",
+            "models": [model.to_dict() for model in model_identifiers],
+        }
         deployment_tempdir = tempfile.mkdtemp()
         zipfile_path = os.path.join(deployment_tempdir, "deployment.zip")
         response = self.session.get_rest_response(
-            url=self.code_deployment_url + f"/{deployment_id}/download",
-            method="GET",
-            contenttype="zip",
+            url=self.deployment_package_url + ":download", data=data, method="POST"
         )
         logging.info("Downloading project deployment archive...")
         with open(zipfile_path, "wb") as f:
@@ -231,8 +190,8 @@ class DeploymentClient:
         require_xai: bool = False,
     ) -> Deployment:
         """
-        Create a deployment for a project, using the /code_deployment endpoint from the
-        backend.
+        Create a deployment for a project, using the /deployment_package:download
+        endpoint from the backend.
 
         :param output_folder: Path to a folder on local disk to which the Deployment
             should be downloaded. If no path is specified, the deployment will not be
@@ -330,36 +289,8 @@ class DeploymentClient:
             for model in sorted_optimized_models
         ]
 
-        # Make the request to prepare the deployment
-        code_deployment = self._request_deployment(model_identifiers=model_identifiers)
-
-        # Wait for the deployment to become available
-        stop_polling = False
-        t_start = time.time()
-        timeout = 600
-        logging.info("Waiting for the deployment to be created...")
-        while not stop_polling and time.time() - t_start < timeout:
-            time.sleep(1)
-            code_deployment = self._get_deployment_status(code_deployment.id)
-            if code_deployment.state == DeploymentState.DONE:
-                stop_polling = True
-            elif code_deployment.state == DeploymentState.FAILED:
-                raise ValueError(
-                    f"The Intel® Geti™ server failed to create deployment for "
-                    f"project '{self.project.name}'."
-                )
-
-        if code_deployment.state != DeploymentState.DONE:
-            raise ValueError(
-                f"The Intel® Geti™ server failed to create deployment for "
-                f"project '{self.project.name}' within 10 minutes. Deployment "
-                f"creation timed out, the current state of the deployment is"
-                f" `{code_deployment.state}` with a progress of "
-                f"{code_deployment.progress:.1f}%"
-            )
-
-        # Fetch the deployment package
-        deployment = self._fetch_deployment(deployment_id=code_deployment.id)
+        # Make the request to prepare and fetch the deployment package
+        deployment = self._fetch_deployment(model_identifiers=model_identifiers)
 
         # Attach the appropriate hyper parameters and model_group_id to the deployed
         # models